refactor(frontend): 工作台子视图移入 views/workspace/ 子目录
