+++ frontend/src/views/WorkspaceView.vue
@@ -25,14 +25,14 @@ import ErasePreview from '@/components/workspace/ErasePreview.vue'
 import OcrPreview from '@/components/workspace/OcrPreview.vue'
 import ImageModal from '@/components/base/ImageModal.vue'
 import ToastContainer from '@/components/base/ToastContainer.vue'
-import Dashboard from '@/views/DashboardView.vue'
-import ReviewView from '@/views/ReviewView.vue'
-import ErrorBank from '@/views/ErrorBankView.vue'
-import ChatView from '@/views/ChatView.vue'
-import SettingsView from '@/views/SettingsView.vue'
-import SplitHistory from '@/views/SplitHistoryView.vue'
-import NoteView from '@/views/NoteView.vue'
-import ChatPage from '@/views/ChatPageView.vue'
+import Dashboard from '@/views/workspace/DashboardView.vue'
+import ReviewView from '@/views/workspace/ReviewView.vue'
+import ErrorBank from '@/views/workspace/ErrorBankView.vue'
+import ChatView from '@/views/workspace/ChatView.vue'
+import SettingsView from '@/views/workspace/SettingsView.vue'
+import SplitHistory from '@/views/workspace/SplitHistoryView.vue'
+import NoteView from '@/views/workspace/NoteView.vue'
+import ChatPage from '@/views/workspace/ChatPageView.vue'
 
 // ---- Composables ----
 const { currentUser } = useAuth()
