+++ frontend/src/views/workspace/ChatPageView.vue
