+++ frontend/src/views/workspace/ChatView.vue
