+++ frontend/src/views/workspace/ErrorBankView.vue
