+++ frontend/src/views/workspace/NoteView.vue
