+++ frontend/src/views/workspace/SplitHistoryView.vue
